requirements.txt
pyserial==3.*
psycopg2-binary==2.*
rich==13.*
hexdump==3.*
click==8.*
